refactor: separate notification routes and controller logic
